+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### Unreleased
+
+* `kind` field of `RtpHeaderExtension` is no longer optional.
+* Refactor API inconsistency in internal RTP Observer communication with worker.
 
 ### 3.7.14